# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "API Gateway Service API",
        "description": "",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "basePath": "/",
...
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/users/{id}/wallet": {
            "post": {
                "tags": ["Users"],
                "summary": "Creates a new wallet associated to user",
                "description": "Returns created wallet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A wallet",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string",
                                    "format":"uuid"
                                },
                                "address": {
                                    "type": "string"
                                },
                                "privateKey": {
                                    "type": "string"
                                },
                                "balance": {
                                    "type": "number"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "get": {
                "tags": ["Users"],
                "summary": "Get wallet by user id",
                "description": "Returns a wallet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true
                    },
                    {
                        "name": "authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A wallet",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string",
                                    "format":"uuid"
                                },
                                "address": {
                                    "type": "string"
                                },
                                "privateKey": {
                                    "type": "string"
                                },
                                "balance": {
                                    "type": "number"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/{id}/deposit": {
            "post": {
                "tags": ["Users"],
                "summary": "Makes a new deposit",
                "description": "Returns deposit receipt",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "deposit",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "required": [
                                "amount"
                            ],
                            "properties": {
                                "amount": {
                                    "type": "string"
                                }
                            }
                        }
                    },   
                    {
                        "name": "authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A deposit receipt",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string",
                                    "format": "uuid"
                                },
                                "sender_address": {
                                    "type": "string",
                                    "format": "uuid"
                                },
                                "amount_sent": {
                                    "type": "number"
                                },
                                "tx_hash": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
}